requirements.txt
-e .
pytest
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_sns
aws-cdk.aws_sqs
aws-cdk.aws_sns_subscriptions
aws-cdk.aws_iam

